requirements.txt
robyn
requests
PyYAML
psutil
Jinja2
APScheduler
Nuitka